update: 更新接口规范

+++ src/main/java/edu/tust/neusoft/backend/repository/WalletRepository.java
@@ -4,5 +4,5 @@ import edu.tust.neusoft.backend.model.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WalletRepository extends JpaRepository<Wallet, Integer> {
-    Wallet findByUserId(int userId);
+    Wallet findByUserId(Long userId);
 }
